--- app/app.py
@@ -1,3 +1,0 @@
-'''
-Создаем приложение для инференса
-'''
