+++ main.py
@@ -1,0 +1,6 @@
+'''
+Достаем предобученную модель и запускаем приложение с ней
+'''
+
+if __name__ == "__main__":
+    pass
