--- main/main.py
@@ -1,6 +1,0 @@
-'''
-Достаем предобученную модель и запускаем приложение с ней
-'''
-
-if __name__ == "__main__":
-    pass
