--- models/pretrained/config.json
@@ -1,12 +1,0 @@
-{
-    "model_type": "Transformer",
-    "max_seq_len": 300,
-    "model_params": {
-        "vocab_size": 69715,
-        "embed_dim": 128,
-        "num_heads": 4,
-        "num_layers": 2,
-        "num_classes": 2,
-        "max_seq_len": 300
-    }
-}
