+++ pyproject.toml
@@ -1,0 +1,31 @@
+[tool.poetry]
+name = "monkey-coding-dl-project"
+version = "0.1.0"
+description = "HSE DL spring 2025 project"
+authors = [
+    "Michael Litvinov",
+    "Kamil Gabidullin",
+]
+readme = "README.md"
+packages = [
+    { include = "src" },
+]
+
+[tool.poetry.dependencies]
+python = "^3.12"
+datasets = "3.6.0"
+matplotlib = "3.10.3"
+nltk = "3.9.1"
+numpy = "2.3.0"
+pandas = "2.3.0"
+scikit-learn = "1.7.0"
+seaborn = "0.13.2"
+torch = "2.7.1"
+transformers = "4.52.4"
+jupyter = "^1.1.1"
+ipykernel = "^6.29.5"
+
+
+[build-system]
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
