--- requirements.txt 
@@ -1,80 +1,0 @@
-aiohappyeyeballs==2.6.1
-aiohttp==3.12.12
-aiosignal==1.3.2
-appnope==0.1.4
-asttokens==3.0.0
-attrs==25.3.0
-certifi==2025.4.26
-charset-normalizer==3.4.2
-click==8.2.1
-comm==0.2.2
-contourpy==1.3.2
-cycler==0.12.1
-datasets==3.6.0
-debugpy==1.8.14
-decorator==5.2.1
-dill==0.3.8
-executing==2.2.0
-filelock==3.18.0
-fonttools==4.58.2
-frozenlist==1.7.0
-fsspec==2025.3.0
-hf-xet==1.1.3
-huggingface-hub==0.32.5
-idna==3.10
-ipykernel==6.29.5
-ipython==9.3.0
-ipython_pygments_lexers==1.1.1
-jedi==0.19.2
-Jinja2==3.1.6
-joblib==1.5.1
-jupyter_client==8.6.3
-jupyter_core==5.8.1
-kiwisolver==1.4.8
-MarkupSafe==3.0.2
-matplotlib==3.10.3
-matplotlib-inline==0.1.7
-mpmath==1.3.0
-multidict==6.4.4
-multiprocess==0.70.16
-nest-asyncio==1.6.0
-networkx==3.5
-nltk==3.9.1
-numpy==2.3.0
-packaging==25.0
-pandas==2.3.0
-parso==0.8.4
-pexpect==4.9.0
-pillow==11.2.1
-platformdirs==4.3.8
-prompt_toolkit==3.0.51
-propcache==0.3.2
-psutil==7.0.0
-ptyprocess==0.7.0
-pure_eval==0.2.3
-pyarrow==20.0.0
-Pygments==2.19.1
-pyparsing==3.2.3
-python-dateutil==2.9.0.post0
-pytz==2025.2
-PyYAML==6.0.2
-pyzmq==26.4.0
-regex==2024.11.6
-requests==2.32.4
-scikit-learn==1.7.0
-scipy==1.15.3
-setuptools==80.9.0
-six==1.17.0
-stack-data==0.6.3
-sympy==1.14.0
-threadpoolctl==3.6.0
-torch==2.7.1
-tornado==6.5.1
-tqdm==4.67.1
-traitlets==5.14.3
-typing_extensions==4.14.0
-tzdata==2025.2
-urllib3==2.4.0
-wcwidth==0.2.13
-xxhash==3.5.0
-yarl==1.20.1
