+++ src/app/app.py
@@ -1,0 +1,3 @@
+'''
+Создаем приложение для инференса
+'''
